requirements.txt
arxiv==2.1.3
azure_storage==0.37.0
bert_score==0.3.13
data_loading==0.0.1
diskcache==5.6.3
faiss==1.5.3
llama_index==0.12.25
nltk==3.9.1
numpy==2.2.4
pandas==2.2.3
pydantic==2.10.6
python-dotenv==1.0.1
Requests==2.32.3
rerankers==0.9.1
rouge_score==0.1.2
semanticscholar==0.10.0
torch==2.6.0
tqdm==4.67.1
transformers==4.50.0.dev0
